# Empresa-OpenApi.yaml
openapi: 3.0.0
info:
  title: Cuentas
  version: '1.0'
  description: Operaciones con Cuentas
  contact:
    name: William
    url: 'http://wquimis.com'
    email: wquimis@gmail.com
  termsOfService: 'http://wquimis.com'
servers:
  - url: 'http://localhost:3000'
...
paths:
...
  /cuentas:
    parameters: []
    post:
      summary: Crear Cuenta
      operationId: crearcuenta
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cuenta'
              examples:
                cuenta:
                  value:
                    idCuenta: 0
                    idCliente: 0
                    numeroCuenta: '000000'
                    tipoCuenta: A
                    saldoInicial: 0
                    estado: A
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cuenta'
            examples:
              cuenta:
                value:
                  idCuenta: 0
                  idCliente: 0
                  numeroCuenta: '000000'
                  tipoCuenta: A
                  saldoInicial: 0
                  estado: A
        description: Devuelve el objeto creado con su ID
      tags:
        - Cuentas
      description: Creacion de cuentas
    get:
      summary: Consultar Cuentas
      operationId: obtenercuenta
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cuenta'
              examples:
                cuenta:
                  value:
                    - idCuenta: 0
                      idCliente: 0
                      numeroCuenta: '000000'
                      tipoCuenta: A
                      saldoInicial: 0
                      estado: A
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      description: Consulta de cuentas
      parameters:
        - schema:
            type: integer
          in: query
          name: idcuenta
          description: Identificador de cuenta
      tags:
        - Cuentas
  '/cuentas/{idcuenta}':
    put:
      summary: Actualizar Cuenta
      operationId: actualizarcuenta
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cuenta'
              examples:
                cuenta:
                  value:
                    idCuenta: 0
                    idCliente: 0
                    numeroCuenta: '000000'
                    tipoCuenta: A
                    saldoInicial: 0
                    estado: A
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cuenta'
            examples:
              cuenta:
                value:
                  idCuenta: 0
                  idCliente: 0
                  numeroCuenta: '000000'
                  tipoCuenta: A
                  saldoInicial: 0
                  estado: A
          application/xml:
            schema:
              type: object
              properties: {}
          multipart/form-data:
            schema:
              type: object
              properties: {}
        description: Objeto a actualizar
      tags:
        - Cuentas
      description: Actualizar Cuenta
    parameters:
      - schema:
          type: integer
        name: idcuenta
        in: path
        description: Identificador de la cuenta
        required: true
    delete:
      summary: Eliminar Cuenta
      operationId: eliminarcuenta
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      tags:
        - Cuentas
      description: Elimina cuenta
...
components:
  schemas:
    Cuenta:
      title: Cuenta
      type: object
      x-examples: {}
      description: Modelo de cuenta
      properties:
        idCuenta:
          type: integer
        idCliente:
          type: integer
        numeroCuenta:
          type: string
          minLength: 6
          maxLength: 10
          default: '000000'
        tipoCuenta:
          type: string
          maxLength: 1
          default: A
        saldoInicial:
          type: number
          maximum: 999999999994.999
          minimum: -999999999999.999
          default: 0
          format: double
        estado:
          type: string
          maxLength: 1
          default: A
      required:
        - idCliente
        - numeroCuenta
        - tipoCuenta
        - saldoInicial
    Error:
      title: Error
      type: object
      x-examples: {}
      description: Modelos de errores
      properties:
        idError:
          type: integer
        mensajeTecnico:
          type: string
        mensajeUsuario:
          type: string
        nivelError:
          type: string
...